convert Panel to ts
